Store driver version in a generated class
Accessing the driver version from META-INF/MANIFEST.MF has proved to be
too fragile, as it fails when the driver is deployed in ways in which
that file is difficult to access or no longer even exists.

Generating a class at build time and storing it in the class path will
be robust against any attempts to re-package.

JAVARS-106